[Builtins] Quit using 'SomeConstantOf'
