+++ plutus-core/plutus-core/examples/PlutusCore/Examples/Builtins.hs
@@ -18,6 +18,7 @@ module PlutusCore.Examples.Builtins where
 
 import           PlutusCore
 import           PlutusCore.Constant
+import           PlutusCore.Constant.Debug
 import           PlutusCore.Evaluation.Machine.ExBudget
 import           PlutusCore.Evaluation.Machine.Exception
 import           PlutusCore.Pretty
