+++ plutus-core/plutus-core/src/PlutusCore/Evaluation/Result.hs
@@ -87,6 +87,9 @@ instance Alternative EvaluationResult where
     EvaluationSuccess x <|> _ = EvaluationSuccess x
     EvaluationFailure   <|> a = a
 
+instance MonadFail EvaluationResult where
+    fail _ = EvaluationFailure
+
 instance PrettyBy config a => PrettyBy config (EvaluationResult a) where
     prettyBy config (EvaluationSuccess x) = prettyBy config x
     prettyBy _      EvaluationFailure     = "Failure"
